[김선유] 댓글 추가, 삭제 구현. board file 추가 구현중.. view 구현 필요 mapper - board 파일 추가, 삭제 구현중, 댓글 추가 삭제 구현.
service - board 파일 추가, 삭제 구현중, 댓글 추가 삭제 구현.

controller - board 파일 추가, 삭제 구현중.. 댓글 추가 삭제 구현.

view - 댓글 구현 완료, 게시판 파일 추가 구현 필요.

+++ OBO/src/main/java/com/gd/obo/vo/BoardFile.java
@@ -9,6 +9,6 @@ public class BoardFile {
 	String createDate;
 	String boardFileOriginalName;
 	String boardFileName;
-	int boardFileSize;
+	double boardFileSize;
 	String boardFileExt;
 }
